+++ plugins/.gitignore
@@ -1,2 +1,3 @@
-*
+/*
 !.gitignore
+!autocomplete
