+++ script.js
@@ -1,5 +1,4 @@
 import { initialize } from '/javascripts/radiata'
 import '/javascripts/gallery'
-import '/javascripts/autocomplete'
 
 initialize()
